Embed Metadata
Using an embedded structure brings the benefit of providing a new
GetMetadata method for embedding structs, which itself can than
be used for interfaces. As the etag attribute might be used for
caching data this will make implementation of a caching machism
easier.

While most of the changes are non breaking, 2 objects included
Metadata not as pointer, Location and Snapshots. Metadata in
location is likely based on a bug in the swagger doc as
no metadata is delivered & and documentation also does not
mention metadata for location.