Make firewall rules updatable
For Ionos Cloud API `null` values have a meaning in the context of
firewalls. E.g. when editing a rule to remove a MAC address restriction,
the `sourceMac` JSON field must be set to `null`. So when marshaling the
data, the field must not just be omitted when it's `nil`.

So this change removes most of the `omitempty` options in the JSON
struct tags. However, not so for the `Protocol` field. It must not be
sent in updates, so `omitempty` is required there.